requirements.txt
redis
fastapi
uvicorn[standard]
websocket
load_dotenv
pydantic